Fix x86 CI tests for native Int and OverflowError fuzz.
- Compare maketerm/codegen against Int / Vector{Int}, not Int64
- Treat OverflowError like DomainError in fuzz expression generation
  (Int32 multiply overflows on 32-bit hosts)

+++ test/basics.jl
@@ -1005,7 +1005,7 @@ end
     # Promoted symtype is a subtype of referred
     @syms x::Int y::Int
     new_expr = SymbolicUtils.maketerm(typeof(ref_expr), (+), [x, y], nothing)
-    @test symtype(new_expr) == Int64
+    @test symtype(new_expr) == Int
 
     # Check that the Array type does not get changed to AbstractArray
     new_expr = SymbolicUtils.maketerm(
